requirements.txt
numpy
pyopengl
wxpython
Pillow
pycollada
